refactor(Layout): flatten Layout/, move Stack to own module
Flatten the Layout directory by removing nested subdirectories:
- Move Stack/ to its own top-level module (@xds/core/Stack)
- Pull XDSLayout files up from Layout/XDSLayout/ to Layout/
- Pull container.stylex up from Layout/Container/ to Layout/
- Remove empty Container/ and XDSLayout/ subdirectories

Layout/ now contains only layout-specific files (XDSLayout, container
utility) matching the same flat structure as every other component.
