+++ packages/core/src/Layout/index.ts
@@ -1,20 +1,18 @@
 /**
  * @file index.ts
- * @input Imports layout utilities and components from subfolders
+ * @input Imports layout utilities and components
  * @output Exports XDS layout system
  * @position Entry point for @xds/core/Layout
  *
  * SYNC: When modified, update /packages/core/src/Layout/README.md
  */
 
-// Utilities
-export {container} from './Container/container.stylex';
-export type {
-  ContainerOptions,
-  SpacingToken,
-} from './Container/container.stylex';
+// Container utility
+export {container} from './container.stylex';
+export type {ContainerOptions, SpacingToken} from './container.stylex';
 
-export {stack} from './Stack/stack.stylex';
+// Stack utilities (re-exported from Stack module)
+export {stack} from '../Stack/stack.stylex';
 export type {
   StackOptions,
   StackDirection,
@@ -22,43 +20,43 @@ export type {
   StackMainAlignment,
   StackWrap,
   SpacingScale,
-} from './Stack/stack.stylex';
+} from '../Stack/stack.stylex';
 
-export {stackItem} from './Stack/stackItem.stylex';
+export {stackItem} from '../Stack/stackItem.stylex';
 export type {
   StackItemOptions,
   StackItemCrossAlignSelf,
   StackItemSize,
-} from './Stack/stackItem.stylex';
+} from '../Stack/stackItem.stylex';
 
-// Stack components
-export {XDSHStack, XDSVStack, XDSStackItem} from './Stack';
-export type {XDSHStackProps, XDSVStackProps, XDSStackItemProps} from './Stack';
+// Stack components (re-exported from Stack module)
+export {XDSHStack, XDSVStack, XDSStackItem} from '../Stack';
+export type {XDSHStackProps, XDSVStackProps, XDSStackItemProps} from '../Stack';
 
-// Container components
-export {XDSCard, XDSSection} from './Container';
-export type {
-  XDSCardProps,
-  SizeValue,
-  XDSSectionProps,
-  XDSSectionVariant,
-} from './Container';
+// Container components (re-exported from their own modules)
+export {XDSCard} from '../Card';
+export type {XDSCardProps} from '../Card';
+
+export {XDSSection} from '../Section';
+export type {XDSSectionProps, XDSSectionVariant} from '../Section';
+
+export type {SizeValue} from '../utils/types';
 
 // Layout structure components
-export {
-  XDSLayout,
-  XDSLayoutHeader,
-  XDSLayoutFooter,
-  XDSLayoutContent,
-  XDSLayoutPanel,
-  XDSLayoutAreaContext,
-} from './XDSLayout';
-export type {
-  XDSLayoutProps,
-  XDSLayoutHeight,
-  XDSLayoutHeaderProps,
-  XDSLayoutFooterProps,
-  XDSLayoutContentProps,
-  XDSLayoutPanelProps,
-  LayoutArea,
-} from './XDSLayout';
+export {XDSLayout} from './XDSLayout';
+export type {XDSLayoutProps, XDSLayoutHeight} from './XDSLayout';
+
+export {XDSLayoutHeader} from './XDSLayoutHeader';
+export type {XDSLayoutHeaderProps} from './XDSLayoutHeader';
+
+export {XDSLayoutFooter} from './XDSLayoutFooter';
+export type {XDSLayoutFooterProps} from './XDSLayoutFooter';
+
+export {XDSLayoutContent} from './XDSLayoutContent';
+export type {XDSLayoutContentProps} from './XDSLayoutContent';
+
+export {XDSLayoutPanel} from './XDSLayoutPanel';
+export type {XDSLayoutPanelProps} from './XDSLayoutPanel';
+
+export {XDSLayoutAreaContext} from './XDSLayoutAreaContext';
+export type {LayoutArea} from './XDSLayoutAreaContext';
