+++ packages/core/src/Stack/XDSStackItem.tsx
